redesign ui elements
